+++ tf2onnx/custom_opsets/ms.py
@@ -36,41 +36,3 @@ class Range:
         utils.make_sure(dtype is not None, "Tidx of %s is None", node.name)
         ctx.remove_node(node.name)
         make_range(ctx, node.input[0], node.input[1], node.input[2], node.output[0], node.name, shape, dtype)
-
-
-@tf_op("ReverseSequence", domain=constants.MICROSOFT_DOMAIN)
-class ReverseSequence:
-    @classmethod
-    def version_1(cls, ctx, node, **kwargs):
-        """ReverseSequence"""
-        # T output = ReverseSequence(T input, int32|int64 seq_lengths, @int seq_dim, @int batch_dim)
-        # T output = ReverseSequence(T input, int32 seqence_lens, @int time_axis, @int batch_axis)
-        seq_dim = node.get_attr("seq_dim")
-        utils.make_sure(seq_dim is not None, "sequence dim must be given in {}".format(node.name))
-        seq_dim = seq_dim.i
-        batch_dim = node.get_attr("batch_dim")
-        if batch_dim is not None:
-            batch_dim = batch_dim.i
-        else:
-            batch_dim = 0
-
-        output_dtypes = node.output_dtypes
-        output_shapes = node.output_shapes
-        ctx.remove_node(node.name)
-        node = ctx.make_node(
-            "ReverseSequence",
-            node.input,
-            outputs=node.output,
-            shapes=output_shapes,
-            dtypes=output_dtypes,
-            domain=constants.MICROSOFT_DOMAIN,
-            attr={"time_axis": seq_dim, "batch_axis": batch_dim}
-        )
-
-        seq_len_dtype = ctx.get_dtype(node.input[1])
-        utils.make_sure(seq_len_dtype is not None, "dtype of {} is None".format(node.input[1]))
-        target_dtype = TensorProto.INT32
-        if seq_len_dtype != target_dtype:
-            cast_node = ctx.insert_new_node_on_input(node, "Cast", node.input[1], to=target_dtype)
-            ctx.copy_shape(cast_node.input[0], cast_node.output[0])
-            ctx.set_dtype(cast_node.output[0], target_dtype)
