+++ tf2onnx/onnx_opset/tensor.py
@@ -19,8 +19,10 @@ import tf2onnx
 from tf2onnx import constants, utils
 from tf2onnx.graph_builder import GraphBuilder
 from tf2onnx.handler import tf_op
+from tf2onnx.onnx_opset import nn
 
 logger = logging.getLogger(__name__)
+
 
 # pylint: disable=unused-argument,missing-docstring,unused-variable
 
@@ -775,8 +777,8 @@ class OneHot:
         # in ONNX, op's schema is (input, depth, value, @int axis), meaning of "value" is [off-value, on-value]
         # onnxruntime only supports int64
         output_dtype = ctx.get_dtype(node.input[2])
-        if ctx.is_target(constants.TARGET_RS6)\
-            and output_dtype not in [onnx_pb.TensorProto.INT64, onnx_pb.TensorProto.INT32]:
+        if ctx.is_target(constants.TARGET_RS6) \
+                and output_dtype not in [onnx_pb.TensorProto.INT64, onnx_pb.TensorProto.INT32]:
             logger.warning("unsupported dtype in onnxruntime, onehot-9 can't be used directly")
             cls.version_5(ctx, node, **kwargs)
             return
@@ -792,24 +794,24 @@ class OneHot:
 
         indices = node.input[0]
         if ctx.is_target(constants.TARGET_RS6) \
-           and ctx.get_dtype(indices) != onnx_pb.TensorProto.INT64:
+                and ctx.get_dtype(indices) != onnx_pb.TensorProto.INT64:
             indices = ctx.make_node("Cast", [indices], attr={"to": onnx_pb.TensorProto.INT64}).output[0]
         node.input[0] = indices
 
         if ctx.is_target(constants.TARGET_RS6) \
-           and ctx.get_dtype(depth) != onnx_pb.TensorProto.INT64:
+                and ctx.get_dtype(depth) != onnx_pb.TensorProto.INT64:
             depth = ctx.make_node("Cast", [depth], attr={"to": onnx_pb.TensorProto.INT64}).output[0]
         node.input[1] = depth
 
-        if ctx.is_target(constants.TARGET_RS6)\
-           and output_dtype != onnx_pb.TensorProto.INT64:
+        if ctx.is_target(constants.TARGET_RS6) \
+                and output_dtype != onnx_pb.TensorProto.INT64:
             off_on_value = ctx.make_node("Cast", [off_on_value], attr={"to": onnx_pb.TensorProto.INT64}).output[0]
         node.input[2] = off_on_value
 
         del node.input[3]
 
-        if ctx.is_target(constants.TARGET_RS6)\
-           and output_dtype != onnx_pb.TensorProto.INT64:
+        if ctx.is_target(constants.TARGET_RS6) \
+                and output_dtype != onnx_pb.TensorProto.INT64:
             new_node_name = utils.make_name("onehot_output")
             new_node = ctx.insert_new_node_on_output("Cast", node.output[0], new_node_name, to=output_dtype)
             ctx.set_dtype(new_node.output[0], output_dtype)
@@ -958,3 +960,102 @@ class NonMaxSuppression:
         squeeze_op = ctx.make_node("Squeeze", [slice_op], attr={"axes": [1]})
         ctx.make_node("Cast", inputs=squeeze_op.output, attr={"to": onnx_pb.TensorProto.INT32},
                       name=node.name, outputs=node.output, dtypes=dtypes, shapes=shapes)
+
+
+@tf_op("ReverseSequence")
+class ReverseSequence:
+    @classmethod
+    def version_8(cls, ctx, node, **kwargs):
+        # T output = ReverseSequence(T input, int32|int64 seq_lengths, @int seq_dim, @int batch_dim)
+        # T output = Scan(int64 sequence_lens, variadic initial_state_and_scan_inputs, @graph body,
+        #                 @ints directions,@int num_scan_inputs)
+        seq_dim = node.get_attr("seq_dim")
+        batch_dim = node.get_attr("batch_dim")
+        batch_major = seq_dim.i == 1 and (batch_dim or batch_dim.i == 0)
+        time_major = batch_dim.i == 1 and (seq_dim or seq_dim.i == 0)
+        perm_val = None
+
+        if not batch_major and not time_major:
+            error_msg = "unsupported attributes, seq_dim:{}, batch_dim:{}".format(seq_dim, batch_dim)
+            raise ValueError(error_msg)
+
+        if time_major:
+            old_shape = ctx.get_shape(node.input[0])
+            old_dtype = ctx.get_dtype(node.input[0])
+            perm_val = [1, 0]
+            rank = len(old_shape)
+            utils.make_sure(rank >= 2, "rank of reverse_sequence input {} is at least 2".format(node.input[0]))
+            perm_val += list(range(2, rank))
+            trans_node = ctx.insert_new_node_on_input(node, "Transpose", node.input[0], perm=perm_val)
+            new_shape = nn.spatial_map(old_shape, perm_val)
+            ctx.set_shape(trans_node.output[0], new_shape)
+            ctx.set_dtype(trans_node.output[0], old_dtype)
+
+        # handle batch_major input
+        node.type = "Scan"
+        node.set_attr("num_scan_inputs", 1)
+        input_dtype = ctx.get_dtype(node.input[0])
+        input_shape = ctx.get_shape(node.input[0])
+
+        g = ctx.create_new_graph_with_same_config()
+        g.parent_graph = ctx
+        g.add_graph_input('X', input_dtype, input_shape[2:])
+        g.make_node('Identity', ['X'], outputs=['Y'])
+        g.add_graph_output('Y', input_dtype, input_shape[2:])
+
+        node.set_body_graph_as_attr("body", g)
+        node.set_attr("directions", [1])  # reverse the scan input
+
+        seq_len_dtype = ctx.get_dtype(node.input[1])
+        if seq_len_dtype != onnx_pb.TensorProto.INT64:
+            cast_node = ctx.insert_new_node_on_input(node, "Cast", node.input[1])
+            cast_node.set_attr("to", onnx_pb.TensorProto.INT64)
+            ctx.set_dtype(cast_node.output[0], onnx_pb.TensorProto.INT64)
+            ctx.copy_shape(node.input[1], cast_node.output[0])
+
+        if time_major:
+            # get back to time_major
+            op_name = utils.make_name(node.name)
+            trans_back_node = ctx.insert_new_node_on_output("Transpose", node.output[0],
+                                                            name=op_name, perm=perm_val)
+            ctx.copy_dtype(node.output[0], trans_back_node.output[0])
+
+        tmp = node.input[0]
+        node.input[0] = node.input[1]
+        node.input[1] = tmp
+
+    @classmethod
+    def version_9(cls, ctx, node, **kwargs):
+        # T output = ReverseSequence(T input, int32|int64 seq_lengths, @int seq_dim, @int batch_dim)
+        # we cannot easily construct reverse_sequence equivalence in opset 9, so we will not support it
+        # here. Actually using loops to do that is kind of meaningless since there will be performance
+        # issue there for sure.
+        raise NotImplementedError("ReverseSequence is not supported to convert in OPSET 9,"
+                                  " if possible please try using OPSET 8, or OPSET >=10 instead.")
+
+    @classmethod
+    def version_10(cls, ctx, node, **kwargs):
+        # T output = ReverseSequence(T input, int32|int64 seq_lengths, @int seq_dim, @int batch_dim)
+        # T output = ReverseSequence(T input, int64 sequence_lens, @int time_axis, @int batch_axis)
+        seq_dim = node.get_attr("seq_dim")
+        utils.make_sure(seq_dim is not None, "sequence dim must be given in {}".format(node.name))
+        seq_dim = seq_dim.i
+        batch_dim = node.get_attr("batch_dim")
+        if batch_dim is not None:
+            batch_dim = batch_dim.i
+        else:
+            batch_dim = 0
+
+        ctx.remove_node(node.name)
+        node = ctx.make_node(
+            "ReverseSequence",
+            node.input,
+            outputs=node.output,
+            attr={"batch_axis": batch_dim, "time_axis": seq_dim}
+        )
+
+        seq_len_dtype = ctx.get_dtype(node.input[1])
+        utils.make_sure(seq_len_dtype is not None, "dtype of {} is None".format(node.input[1]))
+        target_dtype = TensorProto.INT64
+        if seq_len_dtype != target_dtype:
+            ctx.insert_new_node_on_input(node, "Cast", node.input[1], to=target_dtype)
